requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
Flask-Assets
Pillow